TODO

+++ migrations/20250212102904_region_incomes.sql
@@ -4,10 +4,11 @@ CREATE TABLE IF NOT EXISTS region_incomes (
     RegionId INTEGER PRIMARY KEY,
     Year INTEGER,
     Quarter INTEGER,
-    Value FLOAT);
+    Value DECIMAL(18,2));
 -- +goose StatementEnd
 
 -- +goose Down
 -- +goose StatementBegin
 DROP TABLE IF EXISTS region_incomes;
 -- +goose StatementEnd
+//TODO: Формат данных не отрицательный
